requirements.txt
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.1
Django==4.2.3
django-cors-headers==4.2.0
django-location-field==2.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
magic-filter==1.0.9
motor==2.5.1
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.7.0
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7.1
rel==0.4.8
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.15
uvicorn==0.20.0
yarl==1.8.2
